+++ rollup.config.dev.js
@@ -5,6 +5,9 @@ import serve from 'rollup-plugin-serve';
 import { terser } from 'rollup-plugin-terser';
 import json from '@rollup/plugin-json';
 import ignore from './rollup-plugins/ignore';
+import { ignoreTextfieldFiles } from './elements/textfield';
+import { ignoreSelectFiles } from './elements/select';
+import { ignoreSwitchFiles } from './elements/switch';
 
 export default {
   input: ['src/boilerplate-card.ts'],
@@ -30,15 +33,7 @@ export default {
       },
     }),
     ignore({
-      files: [
-        require.resolve('@material/mwc-notched-outline/mwc-notched-outline.js'),
-        require.resolve('@material/mwc-ripple/mwc-ripple.js'),
-        require.resolve('@material/mwc-menu/mwc-menu.js'),
-        require.resolve('@material/mwc-menu/mwc-menu-surface.js'),
-        require.resolve('@material/mwc-list/mwc-list.js'),
-        require.resolve('@material/mwc-list/mwc-list-item.js'),
-        require.resolve('@material/mwc-icon/mwc-icon.js'),
-      ],
+      files: [...ignoreTextfieldFiles, ...ignoreSelectFiles, ...ignoreSwitchFiles].map((file) => require.resolve(file)),
     }),
   ],
 };
